feat: [게시글] 유저 관련 기능 연결 및 삭제 권한 연결

+++ src/main/java/com/example/toy/post/service/PostServiceImpl.java
@@ -3,6 +3,7 @@ package com.example.toy.post.service;
 import com.example.toy.common.exception.CustomException;
 import com.example.toy.common.exception.ErrorCode;
 import com.example.toy.common.response.CreatedData;
+import com.example.toy.common.util.SecurityUtil;
 import com.example.toy.post.dto.req.create.PostCreateRequestDto;
 import com.example.toy.post.dto.req.delete.PostDeleteRequestDto;
 import com.example.toy.post.dto.req.read.PostReadAllRequestDto;
@@ -10,23 +11,29 @@ import com.example.toy.post.dto.req.update.PostUpdateRequestDto;
 import com.example.toy.post.dto.res.PostResponseDto;
 import com.example.toy.post.entity.Post;
 import com.example.toy.post.repository.PostRepository;
-import java.util.ArrayList;
-import java.util.List;
+import com.example.toy.user.entity.User;
+import com.example.toy.user.entity.UserRole;
+import com.example.toy.user.repository.UserRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @Service
 @AllArgsConstructor
 public class PostServiceImpl implements PostService {
 
   private final PostRepository postRepository;
+  private final UserRepository userRepository;
 
   /** 게시글 전체 개수 조회 */
   @Override
   public int getTotalCount(PostReadAllRequestDto postReadAllRequestDto) {
     postReadAllRequestDto.validate();
     List<Post> postList = postRepository.searchPost(postReadAllRequestDto);
+    SecurityUtil.getCurrentUsername();
     return postList.size();
   }
 
@@ -75,10 +82,23 @@ public class PostServiceImpl implements PostService {
   @Override
   @Transactional
   public void deletePost(PostDeleteRequestDto postDeleteRequestDto) {
-    postRepository
-        .findById(postDeleteRequestDto.getPostId())
-        .orElseThrow(() -> new CustomException(ErrorCode.POST_NOT_FOUND));
 
-    postRepository.deletePost(postDeleteRequestDto);
+    Post post =
+        postRepository
+            .findById(postDeleteRequestDto.getPostId())
+            .orElseThrow(() -> new CustomException(ErrorCode.POST_NOT_FOUND));
+
+    String username = SecurityUtil.getCurrentUsername();
+
+    User user =
+        userRepository
+            .findByUsername(username)
+            .orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
+
+    if (UserRole.ADMIN.equals(user.getRole()) || post.getCreatorCd().equals(username)) {
+      postRepository.deletePost(postDeleteRequestDto);
+    } else {
+      throw new CustomException(ErrorCode.ACCESS_DENIED);
+    }
   }
 }
